Delete check_validation
